Create PDF-to-TTS.py
pyttsx3 is a Python library for Text-to-speech. It has many functions which will help the machine to communicate with us. It will help the machine to speak to us PyPDF2 will help with the text from the PDF. A Pure-Python library built as a PDF toolkit. It is capable of extracting document information, splitting documents page by page, merging documents page by page etc.