Create library-managment-system.py
Concatenating GUI with a database into a solid management project is currently the most popular project idea when it comes to an interview project or a megaproject. These projects can be desktop mobile or web apps. So, in this article, we will be building a solid GUI-based Library Management System Python Project with source code and a Database that will manage a library store.

Python supports a built-in database ‘sqlite3’ and a GUI library ‘tkinter’, so we don’t need to explicitly install both.

Click here to get all the resources used in this project.

If you want only code, go to the bottom of the page and copy the code with the copy option available there.

Let’s get started. First, let’s discuss the functionalities of the project.

The Library Management System will consist of
1. Admin Panel – where only authorized admins will be logged in to manage the store through ID and password. Also, it includes the change password option.
2. Dashboard – It will include the name of the logged-in admin, date, and clock. Also, it’s the main core of the project as it will be designed for 8 functionalities :
i. Add Books – add books to store
ii. Issue Books – issue books to students, max limit is 3, and the date of the last issue book will be considered while returning
iii. Edit Books – edit details of the books
iv. Return Books – Return the book and check for fine charges
v. Delete Books – delete books from the store
vi. Search Books – details about the book
vii. Show Books – Show the whole book
viii. Log out – Log out from the current admin session

And most important, it is considered here that students and admins are already registered.

This is a database and GUI-based project, hence we will be using Tkinter and SQLite3 libraries from Python. Other than this we may require some like time or SMTP as per the requirement.
Before moving ahead, let’s go for a short intro to these modules:

Tkinter: It is the standard GUI library in Python. It also provides full OOP programming support and it’s easy and fast
SQLite3: SQLite3 can be integrated with Python using the SQLite3 module. It provides an SQL interface compliant.

We don’t have to install these modules as they come into the built-in package.

Note: In the project total of 3 databases we are going to create as per the functionality. So, it is easier to create them visually instead of writing nasty SQL queries. Hence, it is recommended to install SQLite Studio or any other workbench. We can create a database and tables and add rows and columns with desired constraints with ease from such apps.
Link to download SQLite studio: https://sqlitestudio.pl/

Now that we are done with the introduction and how we are gonna actually do this, let’s step towards implementation. We are building this stepwise one-by-one functionality.

We are using object-oriented programming here, as there are more functions that are kind of connected to the dashboard, hence classes work best.