Create GUI_Interface_for_StudentManagementSystems.py
Initialzing the GUI Window

We have initialized the GUI Window by creating the object of TK() as the main. Set the title of the window, The geometry is the size of the window.

Creating the color variables: Creating the background and foreground color variables.

Creating the StringVar and IntVar variables: Stores the name, email, gender, contact, and stream you entered in the form.

Place the components in the main window

We are creating the main frame where we are labeling the window with the title “Student Management System”.We are also creating 2 frames into that i.e left_frame and right frame specifying the color, width, and height of the frame.

Placing components in the left frame

Here we are placing the components like name, contact number, email address, gender, date of birth, and Stream The label function is used for labeling the components and provides formatting. Entry function is used for entering the text.OptionMenu provides the dropdown list. The Button label is used for the data entry we provided.

Placing components in the Right frame

The right frame displays the data that we have entered, deleted, reset, or modified. We have provided the scroll_bar to scroll the details horizontally and vertically.