Create employee-management-system.py
The Employee Management System Project in Python will be a beginner-friendly Python project that will educate students on how to construct console-based Python applications. We will provide you with the whole source code for the Python project so that you can immediately install it on your PC and begin learning.

Employee Management System Python project assists in the automation of manual operations, saving both time and money. This system protects the professional and personal information of employees and the company. This project can be very helpful if you are creating a final-year project in Python.

This Employee Management System in Python with MySQL database is constructed using the Anaconda Spider in the Python Programming Language. This Simple Project was developed using console-based programming and is linked to a MySQL database as the system’s back-end.

To build an Employee Management System in Python with source code we have to connect the MySQL database to our program, for that, we need to install some Packages